+++ module/myz.js
@@ -10,6 +10,7 @@ import { MYZHumanSheet } from "./actor/human-sheet.js";
 import { MYZNpcSheet } from "./actor/npc-sheet.js";
 import { MYZArkSheet } from "./actor/ark-sheet.js";
 import {MYZVehicleSheet} from "./actor/vehicle-sheet.js";
+import {MYZSpaceshipSheet} from "./actor/spaceship-sheet.js";
 import { MYZItem } from "./item/item.js";
 import { MYZItemSheet } from "./item/item-sheet.js";
 import { MYZDieBase } from "./MYZDice.js";
@@ -36,6 +37,7 @@ Hooks.once("init", async function () {
         MYZNpcSheet,
         MYZArkSheet,
         MYZVehicleSheet,
+        MYZSpaceshipSheet,
         rollItemMacro,
         DiceRoller,
         RollDialog,
@@ -112,6 +114,10 @@ Hooks.once("init", async function () {
     });
     Actors.registerSheet("mutant-year-zero", MYZVehicleSheet, {
         types: ["vehicle"],
+        makeDefault: true,
+    });
+    Actors.registerSheet("mutant-year-zero", MYZSpaceshipSheet, {
+        types: ["spaceship"],
         makeDefault: true,
     });
     Items.unregisterSheet("core", ItemSheet);