+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 The Mutant Year Zero system for FoundryVTT
 
+
+## 11.0.4
+
+- Added new actor type: Spaceship.
+- Added Rot Protection value to Armor.
+- Added a button to roll Rot Protection for equipped armor items.
+- On the weapon and armor sheet, the artifact stats are moved under the "Artifact" tab.
+- Changed all text editors to Prose Mirror.
+- Added default images for different actor types when a new actor is created.
+- Added icons for the roll/push macros.
+- Fixed stunt description for the cerature that has a skill from other creature types
+
+
 ## 11.0.3
 
 - Roll window improvement (all modifiers visible in the dialog)